major-refactor(TransactionTab): Moved TransactionsTab component from flat files to a structured directory with separated concerns, seperated hooks, types and UI logic
